feat: improve navbar loading state
