--- lefthook.yml
@@ -1,7 +1,0 @@
-pre-commit:
-  parallel: true
-  commands:
-    format:
-      glob: "*.{ts,tsx,js,jsx,mdx}"
-      run: pnpm format -- {staged_files}
-      stage_fixed: true
