+++ package.json
@@ -6,8 +6,7 @@
     "build": "turbo build",
     "dev": "turbo dev",
     "lint": "turbo lint",
-    "format": "turbo run format --continue -- --write --cache --cache-location .cache/.prettiercache",
-    "prepare": "lefthook install"
+    "format": "turbo run format --continue -- --write --cache --cache-location .cache/.prettiercache"
   },
   "devDependencies": {
     "@workspace/eslint-config": "workspace:*",
@@ -15,8 +14,7 @@
     "prettier": "catalog:",
     "prettier-plugin-tailwindcss": "^0.6.14",
     "turbo": "^2.5.6",
-    "typescript": "catalog:",
-    "lefthook": "^1.12.3"
+    "typescript": "catalog:"
   },
   "packageManager": "pnpm@10.15.0",
   "engines": {
