+++ projects/app/src/components/user-avatar.tsx
@@ -15,9 +15,10 @@ import { Sparkles } from 'lucide-react';
 import { useRouter } from 'next/navigation';
 import { useCustomer } from 'autumn-js/react';
 import CheckoutDialog from './autumn/checkout-dialog';
+import { Skeleton } from '@workspace/ui/components/skeleton';
 
 export function UserAvatar() {
-  const { data: session } = authClient.useSession();
+  const { data: session, isPending } = authClient.useSession();
   const { checkout, check } = useCustomer();
   const router = useRouter();
 
@@ -35,15 +36,19 @@ export function UserAvatar() {
     <DropdownMenu>
       <DropdownMenuTrigger asChild>
         <Avatar className="hover:cursor-pointer hover:opacity-80">
-          <AvatarImage
-            src={session?.user.image ?? 'https://github.com/shadcn.png'}
-            alt={session?.user.name ?? 'User avatar'}
-          />
-          <AvatarFallback>
-            {session?.user.name.charAt(0).toUpperCase() ??
-              session?.user.email.charAt(0).toUpperCase() ??
-              'U'}
-          </AvatarFallback>
+          {isPending ? (
+            <Skeleton className="h-9 w-9 rounded-full" />
+          ) : (
+            <>
+              <AvatarImage
+                src={session?.user.image ?? 'https://github.com/shadcn.png'}
+                alt={session?.user.name ?? 'User avatar'}
+              />
+              <AvatarFallback>
+                {session?.user.name.charAt(0).toUpperCase() ?? 'U'}
+              </AvatarFallback>
+            </>
+          )}
         </Avatar>
       </DropdownMenuTrigger>
       <DropdownMenuContent>
